Remove external prosemirror example configs
