+++ app/javascript/components/config/nodes.js
@@ -1,11 +1,126 @@
-// import { nodes } from 'prosemirror-schema-basic'
-import { nodes } from './schema-basic'
-
-import { orderedList, bulletList, listItem } from 'prosemirror-schema-list'
-// import { orderedList, bulletList, listItem } from './schema-list'
-
+import { orderedList, bulletList, listItem } from './schema-list'
 import { tableNodes } from 'prosemirror-tables'
 import { footnoteNodes } from '@aeaton/prosemirror-footnotes'
+
+const pDOM = ["p", 0], blockquoteDOM = ["blockquote", 0], hrDOM = ["hr"],
+      preDOM = ["pre", ["code", 0]], brDOM = ["br"]
+
+const nodes = {
+  header: {
+    attrs: {level: {default: 0}},
+    content: "inline*",
+    group: "block",
+    defining: true,
+    parseDOM: [{tag: "head", preserveWhitespace: false}],
+    // toDOM() { return [{tag: "h7", attrs: {level: 1}}]}
+    toDOM() { return ["p", 0] }
+  },
+
+  // content: {
+  //   content: "inline*",
+  //   group: "block",
+  //   parseDom: [{tag: "<__content__>"}],
+  //   toDOM() { return ["p", 0]}
+  // },
+
+  // :: NodeSpec The top level document node.
+  doc: {
+    content: "block+"
+  },
+
+  // :: NodeSpec A plain paragraph textblock. Represented in the DOM
+  // as a `<p>` element.
+  paragraph: {
+    content: "inline*",
+    group: "block",
+    parseDOM: [{tag: "p"}],
+    toDOM() { return pDOM }
+  },
+
+  // :: NodeSpec A blockquote (`<blockquote>`) wrapping one or more blocks.
+  blockquote: {
+    content: "block+",
+    group: "block",
+    defining: true,
+    parseDOM: [{tag: "blockquote"}],
+    toDOM() { return blockquoteDOM }
+  },
+
+  // :: NodeSpec A horizontal rule (`<hr>`).
+  horizontal_rule: {
+    group: "block",
+    parseDOM: [{tag: "hr"}],
+    toDOM() { return hrDOM }
+  },
+
+  // :: NodeSpec A heading textblock, with a `level` attribute that
+  // should hold the number 1 to 6. Parsed and serialized as `<h1>` to
+  // `<h6>` elements.
+  heading: {
+    attrs: {level: {default: 1}},
+    content: "inline*",
+    group: "block",
+    defining: true,
+    parseDOM: [{tag: "h1", attrs: {level: 1}},
+               {tag: "h2", attrs: {level: 2}},
+               {tag: "h3", attrs: {level: 3}},
+               {tag: "h4", attrs: {level: 4}},
+               {tag: "h5", attrs: {level: 5}},
+               {tag: "h6", attrs: {level: 6}}],
+    toDOM(node) { return ["h" + node.attrs.level, 0] }
+  },
+
+  // :: NodeSpec A code listing. Disallows marks or non-text inline
+  // nodes by default. Represented as a `<pre>` element with a
+  // `<code>` element inside of it.
+  code_block: {
+    content: "text*",
+    marks: "",
+    group: "block",
+    code: true,
+    defining: true,
+    parseDOM: [{tag: "pre", preserveWhitespace: "full"}],
+    toDOM() { return preDOM }
+  },
+
+  // :: NodeSpec The text node.
+  text: {
+    group: "inline"
+  },
+
+  // :: NodeSpec An inline image (`<img>`) node. Supports `src`,
+  // `alt`, and `href` attributes. The latter two default to the empty
+  // string.
+  image: {
+    inline: true,
+    attrs: {
+      src: {},
+      alt: {default: null},
+      title: {default: null}
+    },
+    group: "inline",
+    draggable: true,
+    parseDOM: [{tag: "img[src]", getAttrs(dom) {
+      return {
+        src: dom.getAttribute("src"),
+        title: dom.getAttribute("title"),
+        alt: dom.getAttribute("alt")
+      }
+    }}],
+    toDOM(node) { let {src, alt, title} = node.attrs; return ["img", {src, alt, title}] }
+  },
+
+  // :: NodeSpec A hard line break, represented in the DOM as `<br>`.
+  hard_break: {
+    inline: true,
+    group: "inline",
+    selectable: false,
+    parseDOM: [{tag: "br"}],
+    toDOM() { return brDOM }
+  }
+}
+
+const emDOM = ["em", 0], strongDOM = ["strong", 0], codeDOM = ["code", 0]
 
 const listNodes = {
   ordered_list: {
