+++ package.json
@@ -35,11 +35,8 @@
     "popper.js": "^1.16.0",
     "prop-types": "^15.7.2",
     "prosemirror-dropcursor": "^1.3.2",
-    "prosemirror-example-setup": "^1.0.1",
     "prosemirror-gapcursor": "^1.1.3",
     "prosemirror-model": "^1.7.4",
-    "prosemirror-schema-basic": "^1.0.1",
-    "prosemirror-schema-list": "^1.0.4",
     "prosemirror-state": "^1.2.4",
     "prosemirror-tables": "^1.0.0",
     "prosemirror-view": "^1.12.0",
